// frontend/src/api/generated/schema.ts
export interface components {
    schemas: {
        CreateStudyRequest: {
            title: string;
            description: string;
            /** Format: int32 */
            capacity?: number;
            /** Format: date-time */
            recruitmentDeadline: string;
        };
        ApiResponseStudyResponse: {
            success?: boolean;
            message?: string;
            data?: components["schemas"]["StudyResponse"];
        };
// ...
        ApplicationResponse: {
            /** Format: int64 */
            id?: number;
            /** Format: int64 */
            studyId?: number;
            /** Format: int64 */
            applicantId?: number;
            message?: string;
            /** @enum {string} */
            status?: "PENDING" | "APPROVED" | "REJECTED" | "CANCELED";
            /** Format: date-time */
            decidedAt?: string;
            /** Format: date-time */
            createdAt?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
// ...
    createApplication: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                studyId: number;
            };
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["CreateApplicationRequest"];
            };
        };
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "*/*": components["schemas"]["ApiResponseApplicationResponse"];
                };
            };
        };
    };
// ...
    getMyApplications: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "*/*": components["schemas"]["ApiResponseListApplicationListResponse"];
                };
            };
        };
    };
}